example/requirements.txt
alabaster==0.7.16
asgiref==3.7.2
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.3
django-debug-toolbar==4.3.0
django-modeltranslation==0.18.11
docutils==0.20.1
idna==3.6
imagesize==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
Pygments==2.17.2
requests==2.31.0
ruff==0.3.3
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
